// clients/asset-client/src/types/phongTro/phongTro.d.ts
import { EntityType, SearchBase } from "../general";

export interface PhongTroType extends EntityType {
  chuTroId?: string;
	ngayTrong?: Date;
	ngayBatDau?: Date;
	ngayHetHan?: Date;
	ngayDayTin?: Date;
	ngayDuyet?: Date;
	tang?: number;
	soNguoiOToiDa?: number;
	soPhongNgu?: number;
	soPhongTam?: number;
	trangThai: number;
	luotXem: number;
	goiTin: number;
	soLuotDayTin: number;
	trangThaiDuyet: number;
	dienTich?: number;
	gioGiacTuDo?: boolean;
	coMayGiat?: boolean;
	coDieuHoa?: boolean;
	coNongLanh?: boolean;
	coTuLanh?: boolean;
	coGiuongTu?: boolean;
	coKeBep?: boolean;
	isNoiBat: boolean;
	coBanCong?: boolean;
	coThangMay?: boolean;
	coChoDeXe?: boolean;
	coKhoaVanTay?: boolean;
	khongChungChu?: boolean;
	choNuoiThuCung?: boolean;
	giaChoThue?: number;
	tienCoc?: number;
	giaDien?: number;
	giaNuoc?: number;
	giaInternet?: number;
	giaDichVuChung?: number;
	giaGuiXe?: number;
	giaVeSinh?: number;
	phiDangTin?: number;
	tieuDe: string;
	maPhong?: string;
	tenPhong?: string;
	loaiPhong?: string;
	diaChi?: string;
	maTinh?: string;
	moTa?: string;
	quyDinh?: string;
	lyDoTuChoi?: string;
	nguoiDuyet?: string;
	hinhAnhDaiDien?: string;
	danhSachHinhAnh?: string;
	videoLink?: string;
	fileDinhKemIds?: string[];
	avatarFileId?: string;
	tenLienHe?: string;
	soDienThoaiLienHe?: string;
	zaloLienHe?: string;
	donViDien?: string;
	donViNuoc?: string;
	donViInternet?: string;
	donViDichVuChung?: string;
	quyDinhGioGiac?: string;
	tienNghiKhac?: string;
	tenTinh?: string;
	maHuyen?: string;
	tenHuyen?: string;
	maXa?: string;
	tenXa?: string;
	toaDo?: string;
}

export interface PhongTroCreateOrUpdateType {
  id?: string;
  tieuDe: string;
  maPhong?: string;
  tenPhong?: string;
  loaiPhong?: string;
  diaChi?: string;
  maTinh?: string;
  tenTinh?: string;
  maHuyen?: string;
  tenHuyen?: string;
  maXa?: string;
  tenXa?: string;
  toaDo?: string;
  dienTich?: number;
  giaChoThue?: number;
  tienCoc?: number;
  tang?: number;
  soNguoiOToiDa?: number;
  soPhongNgu?: number;
  soPhongTam?: number;
  giaDien?: number;
  donViDien?: string;
  giaNuoc?: number;
  donViNuoc?: string;
  giaInternet?: number;
  donViInternet?: string;
  giaDichVuChung?: number;
  donViDichVuChung?: string;
  giaGuiXe?: number;
  giaVeSinh?: number;
  gioGiacTuDo?: boolean;
  quyDinhGioGiac?: string;
  coMayGiat?: boolean;
  coDieuHoa?: boolean;
  coNongLanh?: boolean;
  coTuLanh?: boolean;
  coGiuongTu?: boolean;
  coBanCong?: boolean;
  coThangMay?: boolean;
  coChoDeXe?: boolean;
  coKhoaVanTay?: boolean;
  khongChungChu?: boolean;
  choNuoiThuCung?: boolean;
  coKeBep?: boolean;
  tienNghiKhac?: string;
  hinhAnhDaiDien?: string;
  danhSachHinhAnh?: string;
  videoLink?: string;
  tenLienHe?: string;
  soDienThoaiLienHe?: string;
  zaloLienHe?: string;
  moTa?: string;
  quyDinh?: string;
  trangThai?: number;
  ngayTrong?: Date | string;
  isNoiBat?: boolean;
  luotXem?: number;
  chuTroId?: string;
  goiTin?: number;
  ngayBatDau?: Date | string;
  ngayHetHan?: Date | string;
  ngayDayTin?: Date | string;
  soLuotDayTin?: number;
  phiDangTin?: number;
  trangThaiDuyet?: number;
  lyDoTuChoi?: string;
  nguoiDuyet?: string;
  ngayDuyet?: Date | string;
}

export interface PhongTroSearchType extends SearchBase {
  trangThai?: number;
	luotXem?: number;
	goiTin?: number;
	soLuotDayTin?: number;
	trangThaiDuyet?: number;
	isNoiBat?: boolean;
	tieuDe?: string;
}

export interface ThongKeTinDangType {
  tatCa: number;
  dangHienThi: number;
  tinThuong: number;
  tinVip: number;
  choDuyet: number;
  hetHan: number;
  daThue: number;
}